FIXED: Iterating through maximum index of vetor hamming exercise

+++ c/hamming/src/hamming.h
@@ -1,6 +1,6 @@
 #ifndef HAMMING_H
 #define HAMMING_H
 
-size_t compute(const char *lhs, const char *rhs);
+long long compute(const char *lhs, const char *rhs);
 
 #endif
